external/AAAI2022_HCM/requirements.txt
numpy==1.19.4
sklearn_pandas==2.0.3
rpy2==3.4.3
pandas==1.1.5
scipy==1.4.1
astropy==4.2
scikit_learn==0.24.2
